Fixed some bugs with the UserHandler pipeline that allows it to pass the tests

+++ server/src/main/java/service/UserService.java
@@ -15,6 +15,10 @@ public class UserService {
     private final AuthDAO authDAO = new LocalAuthDAO();
 
     public UserResult RegisterUser(UserData uData) throws Exception {
+        if(uData.getUsername()==null||uData.getPassword()==null||uData.getEmail()==null){
+            throw new BadRequestException("Error: bad request");
+        }
+
         UserData userServerData = userDAO.GetUserData(uData.getUsername());
         if (userServerData != null) {
             throw new UserAlreadyExistsException("Error: already taken");
@@ -25,11 +29,13 @@ public class UserService {
     }
 
     public UserResult LoginUser(UserData uData) throws Exception {
-        UserData userServerData = userDAO.GetUserData(uData.getUsername());
-        if (userServerData == null) {
+        if (uData.getUsername()==null||uData.getUsername().isEmpty()||
+                uData.getPassword()==null||uData.getPassword().isEmpty()) {
             throw new BadRequestException("Error: bad request");
         }
-        if (!uData.getPassword().equals(userServerData.getPassword())) {
+        UserData userServerData = userDAO.GetUserData(uData.getUsername());
+
+        if (userServerData == null || !uData.getPassword().equals(userServerData.getPassword())) {
             throw new UserUnauthorizedException("Error: unauthorized");
         }
 
@@ -38,6 +44,9 @@ public class UserService {
     }
 
     public UserResult LogoutUser(String authToken) throws Exception {
+        if(authToken==null){
+            throw new UserUnauthorizedException("Error: unauthorized");
+        }
         String username = authDAO.GetAuthUsername(authToken);
         authDAO.DeleteAuthToken(authToken);
         return new UserResult(200, "", "", "");
